refactor: 그룹 정보 수정 로직 변경
