+++ pages/group/[id]/index.page.tsx
@@ -1,12 +1,12 @@
-import { ReactElement, useState } from "react";
+import { ReactElement } from "react";
 
-import Image from "next/image";
-import Router, { useRouter } from "next/router";
+import { useRouter } from "next/router";
 import { GetServerSidePropsContext } from "next/types";
 
 import { QueryClient, dehydrate } from "@tanstack/react-query";
 import classNames from "classnames/bind";
 
+import ActionButton from "./components/ActionButton";
 import GroupForm from "./components/GroupForm";
 import InviteForm from "./components/InviteForm";
 import MemberList from "./components/MemberList";
@@ -14,14 +14,9 @@ import { getGroupInfo, getUserInfo } from "@/apis/apis";
 import { setContext } from "@/apis/axios";
 import Layout from "@/components/Layout";
 import MetaData from "@/components/MetaData";
-import { MODAL_TYPE } from "@/constants/modalType";
-import { PAGE_PATH } from "@/constants/pagePath";
 import { groupKey, userInfoKey } from "@/constants/queryKey";
-import { useDeleteGroup } from "@/hooks/useDeleteGroup";
-import { useExitGroup } from "@/hooks/useExitGroup";
 import { useGetGroupInfo } from "@/hooks/useGetGroupInfo";
 import { useGetUserInfo } from "@/hooks/useGetUserInfo";
-import { useModalStore } from "@/store/useModalStore";
 
 import styles from "./Group.module.scss";
 
@@ -51,30 +46,11 @@ export const getServerSideProps = async (context: GetServerSidePropsContext) => 
 };
 
 export default function Group() {
-  const [isInfoTextView, setIsInfoTextView] = useState(false);
   const { query } = useRouter();
   const { data: groupInfo } = useGetGroupInfo(Number(query.id));
   const { data: userInfo } = useGetUserInfo();
 
-  const { openModal } = useModalStore();
-
   const isCurrentUserLeader = userInfo?.accountId === groupInfo?.crewMembers[0].accountId;
-
-  const handleInfoTextMouseOver = () => {
-    setIsInfoTextView(true);
-  };
-
-  const handleInfoTextMouseLeave = () => {
-    setIsInfoTextView(false);
-  };
-
-  const handleGroupDeleteButtonClick = () => {
-    openModal(MODAL_TYPE.DELETE_GROUP);
-  };
-
-  const handleGroupExitButtonClick = () => {
-    openModal(MODAL_TYPE.EXIT_GROUP);
-  };
 
   if (!groupInfo || !userInfo) {
     // 추후 로딩 처리로 변경
@@ -93,45 +69,7 @@ export default function Group() {
         <InviteForm inviteeList={groupInfo.waitingMembers} />
         <div className={cn("box")}>
           <GroupForm groupInfo={groupInfo} />
-          <div className={cn("button-box")}>
-            <div
-              onMouseOver={handleInfoTextMouseOver}
-              onMouseLeave={handleInfoTextMouseLeave}
-              className={cn("info-box")}
-            >
-              <Image
-                src="/icons/information-icon.svg"
-                width={40}
-                height={40}
-                alt="그룹 정보 버튼"
-              />
-              {isInfoTextView && (
-                <div className={cn("info-item")}>
-                  <div className={cn("info-text")}>
-                    리더가 그룹을 나가려면 다른 멤버에게 리더를 위임하고 나가야 합니다.
-                    <p className={cn("triangle")} />
-                  </div>
-                </div>
-              )}
-            </div>
-            {isCurrentUserLeader ? (
-              <button
-                type="button"
-                className={cn("group-button")}
-                onClick={handleGroupDeleteButtonClick}
-              >
-                그룹 삭제
-              </button>
-            ) : (
-              <button
-                type="button"
-                className={cn("group-button")}
-                onClick={handleGroupExitButtonClick}
-              >
-                그룹 나가기
-              </button>
-            )}
-          </div>
+          <ActionButton isCurrentUserLeader={isCurrentUserLeader} />
         </div>
       </section>
     </>
